Alpha 1.4 complete Service_Education model

+++ service/service_order/src/main/java/com.gkhy.eduorder/service/OrderService.java
@@ -1,7 +1,8 @@
 package com.gkhy.eduorder.service;
 
 public interface OrderService {
+
     String createOrder(String courseId, String memberIdByJwtToken);
 
-    int count();
+    long count();
 }
